Resolve add/add conflict in zhtp-cli/src/commands/observer.rs between
T3 (PR #2588, development) and T5 (this branch).

Resolution keeps T5's version in every conflicted block because T5 is
the functional superset:
- Import: keeps spawn_crypto (for 128 MB stack protection)
- Key generation: uses spawn_crypto instead of direct ZhtpIdentity::new
- File writes: uses atomic .tmp -> rename pattern with .bak backup
- Startup: cleans stale .tmp files from previous crashed attempts
- Doc comments: documents stack overflow protection rationale

+++ lib-blockchain/src/blockchain/tests/replay_contract_execution_tests.rs
@@ -204,7 +204,7 @@ fn token_creation_tx(
     signer: &PublicKey,
     name: &str,
     symbol: &str,
-    supply: u64,
+    supply: u128,
     treasury_recipient: [u8; 32],
 ) -> Transaction {
     let payload = TokenCreationPayloadV1 {
